# server/swagger.openapi3.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/current-user-sessions": {
      "get": {
        "operationId": "AuthController_getCurrentUserSessions",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Sorting fields in format: [+-][fieldName]"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginationResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "readOnly": true,
                          "nullable": false,
                          "items": {
                            "$ref": "#/components/schemas/SessionResponseDto"
                          }
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      },
      "delete": {
        "operationId": "AuthController_closeCurrentUserSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/current-user-sessions/{id}": {
      "delete": {
        "operationId": "AuthController_closeCurrentUserSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/models-3d/current-user": {
      "get": {
        "operationId": "Model3dController_getCurrentUser3DModels",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categories",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Sorting fields in format: [+-][fieldName]"
            }
          },
          {
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "categories",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginationResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "readOnly": true,
                          "nullable": false,
                          "items": {
                            "$ref": "#/components/schemas/Model3dResponseDto"
                          }
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "models-3d"
        ]
      }
    },
    "/api/models-3d": {
      "get": {
        "operationId": "Model3dController_get3DModels",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categories",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Sorting fields in format: [+-][fieldName]"
            }
          },
          {
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "categories",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginationResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "readOnly": true,
                          "nullable": false,
                          "items": {
                            "$ref": "#/components/schemas/Model3dResponseDto"
                          }
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "models-3d"
        ]
      }
    },
    "/api/models-3d/{modelId}": {
      "get": {
        "operationId": "Model3dController_get3DModel",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "models-3d"
        ]
      },
      "patch": {
        "operationId": "Model3dController_update3DModel",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Model3dUpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "models-3d"
        ]
      },
      "delete": {
        "operationId": "Model3dController_delete3DModel",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "models-3d"
        ]
      }
    },
    "/api/models-3d/{modelId}/save-thumbnail-base64": {
      "post": {
        "operationId": "Model3dController_save3DModelThumbnailFromBase64",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "201": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        },
        "tags": [
          "models-3d"
        ]
      }
    },
    "/api/models-3d/upload": {
      "post": {
        "operationId": "Model3dController_upload3DModel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "201": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        },
        "tags": [
          "models-3d"
        ]
      }
    },
    "/api/models-3d/files/{fileId}/thumbnail": {
      "get": {
        "operationId": "Model3dController_getModels3DThumbnailFile",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "tags": [
          "models-3d"
        ]
      }
    },
    "/api/models-3d/files/{fileId}/{fileName}": {
      "get": {
        "operationId": "Model3dController_getModels3DFile",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "tags": [
          "models-3d"
        ]
      }
    }
  },
  "info": {
    "title": "MeshHub Swagger",
    "description": "API приложения MeshHub",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "x-access-token"
      }
    },
    "schemas": {
...
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "SignupRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "firstName",
          "password",
          "confirmPassword"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ip": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "expireAt": {
            "format": "date-time",
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "ip",
          "createdAt",
          "expireAt"
        ]
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "PaginationDtoSortItem": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "by": {
            "enum": [
              "ASC",
              "DESC"
            ],
            "type": "string"
          }
        },
        "required": [
          "field",
          "by"
        ]
      },
      "PaginationResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "skip": {
            "type": "number",
            "minimum": 0
          },
          "size": {
            "type": "number",
            "minimum": 0
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaginationDtoSortItem"
            }
          },
          "totalCount": {
            "type": "number",
            "minimum": 0
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "skip",
          "size",
          "sort",
          "totalCount",
          "hasMore"
        ]
      },
      "Model3dFileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "extension": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "name",
          "size",
          "extension"
        ]
      },
      "Model3dResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isOwner": {
            "type": "boolean"
          },
          "ownerAvatar": {
            "type": "string"
          },
          "ownerName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "isVisible": {
            "type": "boolean"
          },
          "file": {
            "$ref": "#/components/schemas/Model3dFileResponseDto"
          },
          "description": {
            "type": "object"
          },
          "thumbnail": {
            "type": "string"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryResponse"
            }
          }
        },
        "required": [
          "id",
          "createdAt",
          "isOwner",
          "ownerName",
          "name",
          "isVisible",
          "file"
        ]
      },
      "CategoryRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "Model3dUpdateRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isVisible": {
            "type": "boolean"
          },
          "description": {
            "type": "object"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryRequest"
            }
          }
        }
      }
    }
  }
}